+++ pynautobot/core/endpoint.py
@@ -104,6 +104,7 @@ class Endpoint(object):
             token=self.token,
             http_session=self.api.http_session,
             threading=self.api.threading,
+            max_workers=self.api.max_workers,
             api_version=api_version,
         )
 
@@ -389,14 +390,14 @@ class Endpoint(object):
             api_version=api_version,
         ).options()
         try:
-            post_data = req["actions"]["POST"]
-        except KeyError:
-            raise ValueError("Unexpected format in the OPTIONS response at {}".format(self.url))
-        self._choices = {}
-        for prop in post_data:
-            if "choices" in post_data[prop]:
-                self._choices[prop] = post_data[prop]["choices"]
-
+            post_data = req["schema"]["properties"]
+        except (KeyError, TypeError):
+            raise ValueError(f"Unexpected format in the OPTIONS response at {self.url}")
+        self._choices = {
+            prop: [{"value": x, "display": y} for x, y in zip(post_data[prop]["enum"], post_data[prop]["enumNames"])]
+            for prop in post_data
+            if "enum" in post_data[prop]
+        }
         return self._choices
 
     def count(self, *args, api_version=None, **kwargs):