+++ tests/__init__.py
@@ -5,10 +5,11 @@ import pynautobot
 
 from .util import Response
 
-api = pynautobot.api(
-    "http://localhost:8000",
-    token="abc123",
-)
+with patch("pynautobot.api.version", "2.0"):
+    api = pynautobot.api(
+        "http://localhost:8000",
+        token="abc123",
+    )
 
 HEADERS = {
     "accept": "application/json;",