+++ CHANGELOG.md
@@ -1,5 +1,22 @@
 # Changelog
 
+## v2.0.0
+
+### Significant Updates
+
+New release for nautobot 2.0
+
+- (#135) - Docs Update
+- (#134) - Updates in .choices due to changed `OPTIONS` schema
+- (#130) - Add version constraint on `__init__` to divide 1.X and 2.X.
+
+## 2.0.0-rc.1
+
+### Significant Updates
+
+Release Candidate. 
+New release train for nautobot 2.X
+
 ## v1.5.0
 
 ### New Features